requirements.txt
streamlit
tensorflow
pandas
numpy
matplotlib
opencv-python-headless
Pillow